Fonction deleteLigne est maintenant OK, avec le passage de onSupprimer dans la construction de l'adapter

+++ app/src/main/java/ch/ecoandco/genentes/MaBaseDeDonnee.kt
@@ -190,7 +190,7 @@ class MaBaseDeDonnees(context: Context) : SQLiteOpenHelper(context, "anniversair
         }
     }
 
-    fun deleteLine(idEnfant: Long): Boolean {
+    fun deleteLine(idEnfant: Int): Boolean {
         return try {
             val db = this.writableDatabase
             val rowsAffected = db.delete(
